feat(wizard): Replace recipes with security profiles
This commit replaces the concept of "recipes" with "security profiles" throughout the wizard and documentation to better reflect their purpose.

The new security profiles—Development, Hardened, and Isolated—are based on pre-existing, tested configurations from prebuilt devcontainers, ensuring a more reliable and understandable user experience.

Key changes include:
- Renaming `recipes.ts` to `securityProfiles.ts`
- Updating the wizard to use the new security profiles
- Revising the README to reflect the new terminology and profile details
- Increasing tmpfs sizes for improved performance with read-only filesystems
- Fixing a bug in the git-clone `postCreateCommand` to correctly copy hidden files
